third_party/blink/web_tests/ppapi/plugins/nested-plugin-objects-expected.txt

CONSOLE MESSAGE: Blink Deprecated Test Plugin: Destroying
This tests that objects from plugin objects are properly cleaned up.

Example:
   plugin
     ------- (Creates) ------ Object1
                                  ------- (Creates) ------ Object2

It is important that both Object1 and Object2 cleanup as a result of
cleaning up the plugin.


SUCCESS
